refactor folders naming and more organization
