+++ src/editor/Stage/index.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 import { FC, ReactNode, useRef } from "react";
 import { Stage } from "react-konva";
-import { useEvent, useStyleConfig, useZoom } from "../hooks";
+import { useEvent, useStyleConfig } from "../hooks";
 import { css } from "@stylespixelkit/css";
 import useScreen from "../hooks/screen";
 import { useReference } from "../hooks/reference";
@@ -12,7 +12,6 @@ type Props = {
 };
 
 const PixelKitStage: FC<Props> = ({ children }) => {
-  const { onWheel, zoom: stage } = useZoom();
   const ref = useRef<Konva.Stage>(null);
   useReference({
     type: "STAGE",
@@ -36,10 +35,6 @@ const PixelKitStage: FC<Props> = ({ children }) => {
         justifyContent: "center",
       })}`}
       style={{
-        // background: `url("data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABkAAAAZCAYAAADE6YVjAAAACXBIWXMAAAsTAAALEwEAmpwYAAAAAXNSR0IArs4c6QAAAARnQU1BAACxjwv8YQUAAAAxSURBVHgB7dYxDQAACANBwHbFQ7BAQqd/Ab21Kanju29k9ysMgYCAgICAgICcS8fvGhWuCe+rNdorAAAAAElFTkSuQmCC")`,
-        backgroundPosition: "100% 100%",
-        backgroundAttachment: "fixed",
-        backgroundSize: `calc(10px * ${stage.scale})`,
         backgroundColor: config.backgroundColor,
       }}
     >
@@ -48,14 +43,9 @@ const PixelKitStage: FC<Props> = ({ children }) => {
           ref={ref}
           width={width}
           height={height}
-          onWheel={onWheel}
-          scaleX={stage.scale}
-          scaleY={stage.scale}
           onMouseDown={handleMouseDown}
           onMousemove={handleMouseMove}
           onMouseup={handleMouseUp}
-          x={stage.x}
-          y={stage.y}
         >
           {children}
         </Stage>
