+++ src/editor/hooks/index.ts
@@ -2,4 +2,3 @@ export { default as useSelectedShape } from "./selectedShape/hook";
 export { default as useEvent } from "./events/hook";
 export { default as useStyleConfig } from "./styleConfig/hook";
 export { default as useTool } from "./tool/hook";
-export { default as useZoom } from "./zoom/hook";
