--- src/editor/hooks/zoom/types.ts
@@ -1,4 +1,0 @@
-export type IScrollEvent =
-  | "SCROOLL_VERTICAL"
-  | "SCROOLL_HORIZONTAL"
-  | "SCROOLL_ZOOM";
