+++ src/editor/index.tsx
@@ -1,7 +1,7 @@
-import PixelKitShapes from "./elements";
 import useBrowser from "./hooks/useBrowser/hook";
 import useStopZoom from "./hooks/useStopZoom/hook";
 import LayoutPixelEditor from "./layout";
+import PixelKitShapes from "./shapes";
 import PixelKitStage from "./Stage";
 
 const PixelEditor = () => {
