+++ src/editor/shapes/export.shape.tsx
@@ -1,13 +1,14 @@
 import Konva from "konva";
 import { MutableRefObject, memo, useEffect, useRef, useState } from "react";
-import { Rect, Transformer } from "react-konva";
-import { IShape, IShapeWithEvents } from "./type";
+import { Rect, Text, Transformer } from "react-konva";
+import { IShape, IShapeWithEvents } from "./type.shape";
 import { KonvaEventObject } from "konva/lib/Node";
 import { Html } from "react-konva-utils";
 import { createPortal } from "react-dom";
 import ShapeConfig from "../layout/right/shape/config";
+import { ExportStage } from "../layout/right/export";
 // eslint-disable-next-line react/display-name
-const ShapeBox = memo((item: IShapeWithEvents) => {
+const ShapeExport = memo((item: IShapeWithEvents) => {
   const {
     draggable,
     isSelected,
@@ -147,17 +148,25 @@ const ShapeBox = memo((item: IShapeWithEvents) => {
       >
         {sidebarElement && sidebarElement instanceof Element && isSelected
           ? createPortal(
-              <ShapeConfig
-                id={box.id}
-                tool={box.tool}
-                shadowColor={box.shadowColor}
-                strokeColor={box.stroke}
-                fillColor={`${box.backgroundColor}`}
-                borderRadius={box.borderRadius || 0}
-                onChange={(key, value) => {
-                  handleChangeWithKey(key, value);
-                }}
-              />,
+              <>
+                <ShapeConfig
+                  id={box.id}
+                  tool={box.tool}
+                  shadowColor={box.shadowColor}
+                  strokeColor={box.stroke}
+                  fillColor={`${box.backgroundColor}`}
+                  borderRadius={borderRadius || 0}
+                  onChange={(key, value) => {
+                    handleChangeWithKey(key, value);
+                  }}
+                />
+                <ExportStage
+                  x={box.x}
+                  y={box.y}
+                  height={box.height || 0}
+                  width={box.width || 0}
+                />
+              </>,
               sidebarElement
             )
           : null}
@@ -175,11 +184,15 @@ const ShapeBox = memo((item: IShapeWithEvents) => {
         shadowOffsetY={shadowOffsetY}
         shadowBlur={shadowBlur}
         cornerRadius={borderRadius}
+        dashEnabled
+        fillEnabled={false}
+        dash={[25, 25]}
+        hitStrokeWidth={100}
         fill={backgroundColor}
         ref={shapeRef as MutableRefObject<Konva.Rect>}
         draggable={draggable}
         stroke={stroke}
-        strokeWidth={strokeWidth}
+        strokeWidth={5}
         onClick={shapeClick}
         onDragStart={shapeDragStart}
         onDragMove={shapeDragMove}
@@ -202,10 +215,4 @@ const ShapeBox = memo((item: IShapeWithEvents) => {
   );
 });
 
-export const isPartialBorderRadius = (item: IShapeWithEvents) => {
-  return {
-    cornerRadius: [],
-  };
-};
-
-export default ShapeBox;
+export default ShapeExport;
