+++ src/editor/shapes/line.shape.tsx
@@ -1,10 +1,11 @@
 import Konva from "konva";
-import { MutableRefObject, useEffect, useMemo, useRef } from "react";
+import { MutableRefObject, useEffect, useRef } from "react";
 import { Rect, Transformer } from "react-konva";
-import { IShapeWithEvents } from "../type";
+import { isPartialBorderRadius } from "./box.shape";
+import { IShapeWithEvents } from "./type.shape";
 
-const AtomElementCircle = (item: IShapeWithEvents) => {
-  const { draggable, onChange, rotate, onSelect, isSelected } = item;
+const AtomEditorElementBox = (item: IShapeWithEvents) => {
+  const { rotate, draggable, onChange, onSelect, isSelected } = item;
   const shapeRef = useRef<Konva.Rect>();
   const trRef = useRef<Konva.Transformer>();
 
@@ -17,42 +18,52 @@ const AtomElementCircle = (item: IShapeWithEvents) => {
     }
   }, [isSelected, item, trRef, shapeRef]);
 
-  const borderRadius = useMemo(
-    () => Number(item?.width) + Number(item?.height) / 2,
-    [item]
-  );
-
   return (
     <>
       <Rect
         {...item}
         id={item?.id}
-        key={item.id}
         x={item?.x}
         y={item?.y}
-        name={item.id}
         width={item?.width}
+        fill={item.style?.stroke}
         shadowColor={item?.style?.shadowColor}
         shadowOpacity={item?.style?.shadowOpacity}
         shadowOffsetX={item?.style?.shadowOffset?.x}
         shadowOffsetY={item?.style?.shadowOffset?.y}
         shadowBlur={item?.style?.shadowBlur}
-        height={item?.width}
-        cornerRadius={borderRadius}
-        stroke={item?.style?.stroke}
-        strokeWidth={item?.style?.strokeWidth}
-        rotation={rotate}
-        fill={item?.style?.backgroundColor}
+        height={item?.style?.strokeWidth}
+        cornerRadius={isPartialBorderRadius(item)?.cornerRadius}
         ref={shapeRef as MutableRefObject<Konva.Rect>}
         draggable={draggable}
         onClick={() => onSelect(item)}
         onTap={() => onSelect(item)}
+        rotation={rotate}
         onDragEnd={(e) => {
           onChange({
             ...item,
             x: e.target.x(),
             y: e.target.y(),
           });
+        }}
+        onTransform={(e) => {
+          const rotate = e.target.rotation();
+          if (shapeRef?.current) {
+            1;
+            const node = shapeRef.current;
+            const scaleX = node.scaleX();
+            const scaleY = node.scaleY();
+            node.scaleX(1);
+            node.scaleY(1);
+            onChange({
+              ...item,
+              x: node.x(),
+              y: node.y(),
+              rotate,
+              width: Math.max(5, node.width() * scaleX),
+              height: Math.max(node.height() * scaleY),
+            });
+          }
         }}
         onTransformEnd={(e) => {
           const rotate = e.target.rotation();
@@ -62,7 +73,6 @@ const AtomElementCircle = (item: IShapeWithEvents) => {
             const scaleY = node.scaleY();
             node.scaleX(1);
             node.scaleY(1);
-
             onChange({
               ...item,
               x: node.x(),
@@ -77,9 +87,15 @@ const AtomElementCircle = (item: IShapeWithEvents) => {
       {isSelected && (
         <Transformer
           ref={trRef as MutableRefObject<Konva.Transformer>}
-          keepRatio={true}
+          // keepRatio={false}
+          // enabledAnchors={["middle-right", "middle-left"]}
+          // enabledAnchors={[
+          //   "top-left",
+          //   "top-right",
+          //   "bottom-left",
+          //   "bottom-right",
+          // ]}
           boundBoxFunc={(oldBox, newBox) => {
-            // limit resize
             if (newBox.width < 5 || newBox.height < 5) {
               return oldBox;
             }
@@ -91,4 +107,4 @@ const AtomElementCircle = (item: IShapeWithEvents) => {
   );
 };
 
-export default AtomElementCircle;
+export default AtomEditorElementBox;
